requirements.txt
Flask==2.0.1
tabula-py==2.2.0
pandas==1.3.3
openpyxl==3.0.9
Werkzeug==2.0.1
requests==2.26.0